requirements.txt
cherrypy
pyyaml
pubsub
